backend/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

celery==5.3.6
flower==2.0.1
fastapi==0.109.0
fastapi-pagination==0.12.14
langchain==0.1.3
langchain-community==0.0.15
langchain-core==0.1.15
langchain-openai==0.0.3
loguru==0.7.2
openai==1.9.0
passlib==1.7.4
pinecone-client==3.0.2
pydantic==1.10.10
pydub==0.25.1
python-dotenv==1.0.1
python-jose==3.3.0
redis==5.0.1
python-multipart==0.0.6
cryptography==41.0.5
aiomysql==0.2.0
bcrypt==4.1.2
SQLAlchemy==2.0.23
starlette==0.35.1
uvicorn[standard]==0.27.0
